renamed some files for convention
