+++ client/components/AllProducts.js
@@ -1,9 +1,7 @@
-import React, {Component} from 'react';
-import {connect} from 'react-redux';
-import {fetchProducts} from '../store/allProducts';
-import {fetchCategories} from '../store/allCategories';
-import {fetchCurrentCategory} from '../store/currentCategory';
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
+import { fetchProducts, fetchCategories, fetchCurrentCategory } from '../store';
 
 // Component
 class AllProducts extends Component {
@@ -70,7 +68,6 @@ class AllProducts extends Component {
   )
   }
 }
-
 
 // Container
 const mapState  = ({allProducts, currentUser, allCategories}, ownProps) => {
