+++ client/components/CheckoutForm.js
@@ -1,4 +1,4 @@
-import React, {Component} from 'react';
+import React, { Component } from 'react';
 
 export default class CheckoutForm extends Component {
   constructor(props) {
